refactor(typos): Pull out file logic
