+++ crates/typos/Cargo.toml
@@ -20,11 +20,6 @@ thiserror = "1.0"
 regex = "1.3"
 once_cell = "1.2.0"
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 itertools = "0.9"
-bstr = "0.2"
 log = "0.4"
 unicode-segmentation = "1.7.1"
-derive_more = "0.99.11"
-derive_setters = "0.1"
-content_inspector = "0.2.4"
