+++ crates/typos/src/lib.rs
@@ -1,8 +1,6 @@
 mod dict;
 mod parser;
 
-pub mod checks;
-pub mod report;
 pub mod tokens;
 
 pub use dict::*;
