+++ src/args.rs
@@ -12,13 +12,13 @@ arg_enum! {
     }
 }
 
-pub const PRINT_SILENT: typos::report::PrintSilent = typos::report::PrintSilent;
-pub const PRINT_BRIEF: typos::report::PrintBrief = typos::report::PrintBrief;
-pub const PRINT_LONG: typos::report::PrintLong = typos::report::PrintLong;
-pub const PRINT_JSON: typos::report::PrintJson = typos::report::PrintJson;
+pub const PRINT_SILENT: typos_cli::report::PrintSilent = typos_cli::report::PrintSilent;
+pub const PRINT_BRIEF: typos_cli::report::PrintBrief = typos_cli::report::PrintBrief;
+pub const PRINT_LONG: typos_cli::report::PrintLong = typos_cli::report::PrintLong;
+pub const PRINT_JSON: typos_cli::report::PrintJson = typos_cli::report::PrintJson;
 
 impl Format {
-    pub(crate) fn reporter(self) -> &'static dyn typos::report::Report {
+    pub(crate) fn reporter(self) -> &'static dyn typos_cli::report::Report {
         match self {
             Format::Silent => &PRINT_SILENT,
             Format::Brief => &PRINT_BRIEF,
