+++ src/lib.rs
@@ -1,2 +1,6 @@
+pub mod checks;
 pub mod config;
 pub mod dict;
+pub mod diff;
+pub mod replace;
+pub mod report;
